HW3/requirements.txt
matplotlib==3.7.1
opencv_python==4.7.0.72
Pillow==9.5.0
